Fixed circular dof. Had an error in the weights. Gauss still a bit strange though.
Also played around a bit with the precision of the different color attachments. Not sure why Gauss acts strange at 16-bit precision.

Added some code in the lighting shader to easily render a square, this is nice when debugging the different blurs as you can see how they blur in each axis.